Update requirements.adoc
We received a PR (https://github.com/NetApp/trident/pull/626) to correct the Trident version in the container image list for a patch release. It doesn't seem to be the right thing to do to include a version of this table for every Trident patch release. @juliantap, I instead added instructions to run the tridentctl command to get the correct list of container images for that version of Trident. Let me know what you think.