10925: Use only one auth mode
The user will be allowed to have only one authentication methods(either certificate based or username/password based) in the backend config of ONTAP and Flex group backends.

In case, there is more than one authentication method present, the backend creation will fail with an error that more than one authentication method is present in the config. The user will have to remove one of the authentication methods from the config file.